Move new functions to utils
