+++ package.json
@@ -18,7 +18,6 @@
     "eslint": "^7.13.0",
     "mocha": "^7.2.0",
     "nyc": "^15.1.0",
-    "rewire": "^5.0.0",
     "rimraf": "^3.0.2",
     "typescript": "^4.0.5"
   },
